+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## codem-ffmpeg 0.1.5 (2016/11/25) ##
+
+* Emit a log warning if the "-y" option is not specified, warning of a hang.
+
+## codem-ffmpeg 0.1.4 (2016/11/15) ##
+
+* Added injection of config
+
 ## codem-ffmpeg 0.1.3 (2016/10/31) ##
 
 * Expose a custom `attributes` Map which contains educated guesses about the process